refactor(www): carve out header/ slice + final cleanup
Move header components into header/ slice, delete dead layout.shared,
and update tailwind content globs to final set of slices.
