--- apps/www/lib/layout.shared.tsx
@@ -1,10 +1,0 @@
-import type { BaseLayoutProps } from 'fumadocs-ui/layouts/shared';
-
-export function baseOptions(): BaseLayoutProps {
-  return {
-    nav: {
-      title: 'Ora UI',
-    },
-    githubUrl: 'https://github.com/gregogun/ora-ui',
-  };
-}
