+++ apps/www/tailwind.config.ts
@@ -2,7 +2,7 @@ import type { Config } from 'tailwindcss';
 
 const config: Config = {
   darkMode: 'class',
-  content: ['./{app,components,lib,registry,docs,content,playground}/**/*.{ts,tsx}'],
+  content: ['./{app,docs,playground,registry,header}/**/*.{ts,tsx}'],
 };
 
 export default config;
